ESP32_MQTT/3_Python_MQTT_Client_App/requirements.txt
pysimplegui
AWSIoTPythonSDK
Pillow